Create and design the structure of the calculator
I created the structure of the on-screen calculator using divs and
semantic HTML. There is a display and a keypad of buttons that, right
now, does not work as of the moment.

I utilized Josh W. Comeau's CSS reset to clear the browser's initial
styles. I then designed it using basic CSS styles and the flexblox
layout. The design will be further improved after the functionality of
the calculator is completed.