Add getCalculatorInputs and displayResult function
I added the logic for taking the user's calculator inputs, through the
buttons, and storing them in global variable holders. To determine the
textContent passed by the button, I utilized arrays to efficiently
hold references to the numbers, operation, and command keys. To evaluate
the equation, I used a class constructor that bundles the operator,
first operand, and second operand for the operate function.

Additionally, I added a modulo operation for the operate function.

I also added a separate function that would display the calculator
results into the HTML calculator display by displaying the values inside
the global variable holders.

I also added a class to make the placeholder input of the calculator
display hidden from view until the user inputs a number.