Modify calculator logic to handle edge cases
I modified the calculator logic by handling the following edge cases:

1. To make the calculator only evaluate a single pair of numbers at a
time, I added a logic that checks if all three global variables are not
null, then I run the operate function and assign the result back to the
first operand and then sets the operator to the user's next operator.
This handles the chaining of calculator operations cleanly.

2. To avoid overflowing the display with long decimals, I added the
formatStringPrecision function to only set the value to 10 decimal
places, remove trailing zeroes, and convert the result to string.

3. To replace the displayed result when the user presses a new digit, I
added a logic that would check if the previous expression was evaluated.

4. To make sure that the calculator only runs an operation with one
operator, I added a logic similar to No. 1. I also made sure that if the
user clicks a new operator, it would replace the previous operator.

5. Prevent '=' from running the operate function if not all global
variables are filled with input.

6. Prevent and warn user from dividing by zero.

7. Add a clear button that wipes out all data within the global
variables.

8. Prevent multiple decimal points and multiple zeroes before a decimal
point.

9. Add zero to the beginning of a decimal point if the user did not set
a number before inputting decimal point.

10. Add a clear entry button that removes the last input from the global
variables.