Add keyboard support and modify digit limit
I added a keyboard support for the calculator by utilizing an event
listener in the window and capturing the keydown event to return the
key being pressed. Since my buttons return custom text contents, I
processed some of the keyboard inputs to work with the operation list
and command list inside the processCalculatorInput function.

I also modified the validation for operand digit limit to consider the
decimal point as not being part of the digit limit.