Design calculator based on The Matrix
I designed the calculator based on the movie, The Matrix. I prompted
ChatGPT to give me a color palette that would match the colors used in
The Matrix. I also prompted ChatGPT to make the "code rain" aesthetic
using a script once the DOM is loaded. There is also an easter egg that
randomly shows that the calculator is getting hacked.

I added a 7-segment display font by keshikan and utilized it as the font
for the calculator display. The Github link to his repository containing
the font is attached here: https://github.com/keshikan/DSEG.